refactor: UserRow, UserListPanel을 features/presence로 이동
