+++ src/pages/lobby/LobbyPage.tsx
@@ -1,11 +1,11 @@
 import { useMemo, useState } from 'react'
 import Header from '../../components/Header'
+import { UserListPanel } from '../../features/presence/components/UserListPanel'
 import { useOnlineUsersQuery } from '../../features/presence/hooks'
 import { filterLobbyRooms, type RoomFilter } from './filterRooms'
 import { useLobbyRoomsQuery } from './hooks'
 import { LobbyControls } from './LobbyControls'
 import { RoomGrid } from './RoomGrid'
-import { UserListPanel } from '../../components/UserListPanel'
 
 function LobbyPage() {
   const [searchKeyword, setSearchKeyword] = useState('')
